Include rollups in monthlies failed metric as well as monthlies created from scratch
